Move table properties related utility functions to utils/table-properties.js.
